# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"AddressHiroObject": {
				"properties": {
					"non_fungible_tokens": {},
					"fungible_tokens": {},
					"stx": {
						"properties": {
							"burnchain_unlock_height": {
								"type": "number",
								"format": "double"
							},
							"burnchain_lock_height": {
								"type": "number",
								"format": "double"
							},
							"lock_height": {
								"type": "number",
								"format": "double"
							},
							"locked": {
								"type": "number",
								"format": "double"
							},
							"lock_tx_id": {
								"type": "string"
							},
							"total_received": {
								"type": "number",
								"format": "double"
							},
							"total_sent": {
								"type": "number",
								"format": "double"
							},
							"balance": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"burnchain_unlock_height",
							"burnchain_lock_height",
							"lock_height",
							"locked",
							"lock_tx_id",
							"total_received",
							"total_sent",
							"balance"
						],
						"type": "object"
					}
				},
				"required": [
					"non_fungible_tokens",
					"fungible_tokens",
					"stx"
				],
				"type": "object"
			},
			"AddressMempoolObject": {
				"properties": {
					"mempool_stats": {
						"properties": {
							"tx_count": {
								"type": "number",
								"format": "double"
							},
							"spent_txo_sum": {
								"type": "number",
								"format": "double"
							},
							"spent_txo_count": {
								"type": "number",
								"format": "double"
							},
							"funded_txo_sum": {
								"type": "number",
								"format": "double"
							},
							"funded_txo_count": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"tx_count",
							"spent_txo_sum",
							"spent_txo_count",
							"funded_txo_sum",
							"funded_txo_count"
						],
						"type": "object"
					},
					"chain_stats": {
						"properties": {
							"tx_count": {
								"type": "number",
								"format": "double"
							},
							"spent_txo_sum": {
								"type": "number",
								"format": "double"
							},
							"spent_txo_count": {
								"type": "number",
								"format": "double"
							},
							"funded_txo_sum": {
								"type": "number",
								"format": "double"
							},
							"funded_txo_count": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"tx_count",
							"spent_txo_sum",
							"spent_txo_count",
							"funded_txo_sum",
							"funded_txo_count"
						],
						"type": "object"
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"mempool_stats",
					"chain_stats",
					"address"
				],
				"type": "object"
			},
			"AddressObject": {
				"properties": {
					"btcPubkeySegwit1": {
						"type": "string"
					},
					"btcPubkeySegwit0": {
						"type": "string"
					},
					"ordinalInfo": {
						"$ref": "#/components/schemas/AddressMempoolObject"
					},
					"cardinalInfo": {
						"$ref": "#/components/schemas/AddressMempoolObject"
					},
					"bnsNameInfo": {},
					"stacksTokenInfo": {
						"$ref": "#/components/schemas/AddressHiroObject"
					},
					"stxBalance": {
						"type": "number",
						"format": "double"
					},
					"sBTCBalance": {
						"type": "number",
						"format": "double"
					},
					"ordinal": {
						"type": "string"
					},
					"cardinal": {
						"type": "string"
					},
					"stxAddress": {
						"type": "string"
					}
				},
				"required": [
					"stxBalance",
					"sBTCBalance",
					"ordinal",
					"cardinal",
					"stxAddress"
				],
				"type": "object"
			},
...
			"PSBTHolder": {
				"properties": {
					"txFee": {
						"type": "number",
						"format": "double"
					},
					"b64PSBT": {
						"type": "string"
					},
					"hexPSBT": {
						"type": "string"
					}
				},
				"required": [
					"txFee",
					"b64PSBT",
					"hexPSBT"
				],
				"type": "object"
			},
			"OpReturnRequest": {
				"properties": {
					"feeMultiplier": {
						"type": "number",
						"format": "double"
					},
					"paymentAddress": {
						"type": "string"
					},
					"paymentPublicKey": {
						"type": "string"
					},
					"amountSats": {
						"type": "number",
						"format": "double"
					},
					"signature": {
						"type": "string"
					},
					"recipient": {
						"type": "string"
					},
					"originator": {
						"type": "string"
					}
				},
				"required": [
					"feeMultiplier",
					"paymentAddress",
					"paymentPublicKey",
					"amountSats",
					"recipient",
					"originator"
				],
				"type": "object"
			},
			"Uint8Array": {
				"description": "A typed array of 8-bit unsigned integer values. The contents are initialized to 0. If the\r\nrequested number of bytes could not be allocated an exception is raised.",
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"CommitmentScriptDataType": {
				"properties": {
					"tweakedPubkey": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"tapMerkleRoot": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"tapLeafScript": {},
					"tapInternalKey": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"leaves": {},
					"wsh": {
						"type": "string"
					},
					"witnessScript": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"redeemScript": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"paymentType": {
						"type": "string"
					},
					"script": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"paymentType"
				],
				"type": "object"
			},
			"VoutI": {
				"properties": {
					"value": {
						"type": "number",
						"format": "double"
					},
					"scriptpubkey_address": {
						"type": "string"
					},
					"scriptpubkey_type": {
						"type": "string"
					},
					"scriptpubkey_asm": {
						"type": "string"
					},
					"scriptpubkey": {
						"type": "string"
					}
				},
				"required": [
					"value",
					"scriptpubkey_address",
					"scriptpubkey_type",
					"scriptpubkey_asm",
					"scriptpubkey"
				],
				"type": "object"
			},
			"CommitmentStatus": {
				"enum": [
					0,
					1,
					2,
					3,
					4
				],
				"type": "number"
			},
			"RevealerTxModes": {
				"enum": [
					"OP_RETURN",
					"OP_DROP"
				],
				"type": "string"
			},
			"RevealerTxTypes": {
				"enum": [
					"SBTC_DEPOSIT",
					"SBTC_WITHDRAWAL"
				],
				"type": "string"
			},
			"RevealerTransaction": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"txId": {
						"type": "string"
					},
					"psbt": {
						"type": "string"
					},
					"originator": {
						"type": "string"
					},
					"commitment": {
						"$ref": "#/components/schemas/CommitmentScriptDataType"
					},
					"vout": {
						"$ref": "#/components/schemas/VoutI"
					},
					"signed": {
						"type": "boolean"
					},
					"recipient": {
						"type": "string"
					},
					"amountSats": {
						"type": "number",
						"format": "double"
					},
					"confirmations": {
						"type": "number",
						"format": "double"
					},
					"created": {
						"type": "number",
						"format": "double"
					},
					"updated": {
						"type": "number",
						"format": "double"
					},
					"signature": {
						"type": "string"
					},
					"paymentPublicKey": {
						"type": "string"
					},
					"paymentAddress": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/CommitmentStatus"
					},
					"mode": {
						"$ref": "#/components/schemas/RevealerTxModes"
					},
					"type": {
						"$ref": "#/components/schemas/RevealerTxTypes"
					},
					"blockHeight": {
						"type": "number",
						"format": "double"
					},
					"sbtcPublicKey": {
						"type": "string"
					}
				},
				"required": [
					"txId",
					"originator",
					"signed",
					"recipient",
					"amountSats",
					"confirmations",
					"created",
					"updated",
					"paymentPublicKey",
					"paymentAddress",
					"status",
					"mode",
					"type",
					"blockHeight",
					"sbtcPublicKey"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "revealer-api",
		"version": "1.0.0",
		"description": "Proxy server for ordicall web application to access ordinals and bitcoin servers",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/revealer-api/v1/sbtc/address/balances/{stxAddress}/{cardinal}/{ordinal}": {
			"get": {
				"operationId": "FetchUserBalances",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AddressObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "stxAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "cardinal",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "ordinal",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/revealer-api/v1/sbtc/bitcoin/{txid}/{verbose}": {
			"get": {
				"operationId": "FetchBitcoinTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "verbose",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/revealer-api/v1/sbtc/bitcoin/block/{blockhash}/{verbosity}": {
			"get": {
				"operationId": "FetchBitcoinBlock",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "blockhash",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "verbosity",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/revealer-api/v1/op_return/get-psbt-for-deposit": {
			"post": {
				"operationId": "GetPsbtForDeposit",
				"responses": {
					"200": {
						"description": "unsigned psbt",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PSBTHolder"
								}
							}
						}
					}
				},
				"description": "Build an sBTC deposit PSBT using OP_RETURN for the user to sign and sed.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OpReturnRequest"
							}
						}
					}
				}
			}
		},
		"/revealer-api/v1/op_return/get-psbt-for-withdrawal": {
			"post": {
				"operationId": "GetPsbtForWithdrawal",
				"responses": {
					"200": {
						"description": "unsigned psbt",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PSBTHolder"
								}
							}
						}
					}
				},
				"description": "Build an sBTC withdrawal PSBT using OP_RETURN for the user to sign and sed.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OpReturnRequest"
							}
						}
					}
				}
			}
		},
		"/revealer-api/v1/op_return/broadcast-deposit": {
			"post": {
				"operationId": "SendRawTransaction",
				"responses": {
					"200": {
						"description": "string",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RevealerTransaction"
								}
							}
						}
					}
				},
				"description": "Broadcast signed psbt / transaction. Note the recipient, amount and\npublic key are used to make a temporary unique id. Its possible the user \ncan broadcast the siged psbt outside the bridge but the default case is they\nsig ad broadcast from the bridge app - i which  case we replace the temporary\nunique id with the tx id of the finalised tx. This makes payment lookups a lot simpler",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"maxFeeRate": {
										"type": "number",
										"format": "double"
									},
									"signedPsbtHex": {
										"type": "string"
									},
									"paymentPublicKey": {
										"type": "string"
									},
									"amountSats": {
										"type": "number",
										"format": "double"
									},
									"recipient": {
										"type": "string"
									}
								},
								"required": [
									"maxFeeRate",
									"signedPsbtHex",
									"paymentPublicKey",
									"amountSats",
									"recipient"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/revealer-api/v1/op_return/update-deposit": {
			"post": {
				"operationId": "ClientBroadcastDeposit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RevealerTransaction"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"maxFeeRate": {
										"type": "number",
										"format": "double"
									},
									"signedPsbtHex": {
										"type": "string"
									},
									"paymentPublicKey": {
										"type": "string"
									},
									"amountSats": {
										"type": "number",
										"format": "double"
									},
									"recipient": {
										"type": "string"
									},
									"txId": {
										"type": "string"
									}
								},
								"required": [
									"maxFeeRate",
									"signedPsbtHex",
									"paymentPublicKey",
									"amountSats",
									"recipient",
									"txId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}